+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 Notable changes to Mailpit will be documented in this file.
 
+## [v1.17.0]
+
+### Chore
+- Update caniemail database
+- Update node dependencies
+- Update Go dependencies
+- Auto-rotate thumbnail images based on exif data
+- Replace disintegration/imaging with kovidgoyal/imaging to fix CVE-2023-36308
+- Update API documentation regarding date/time searches & timezones
+- Move Link check & HTML check features out of beta
+- Remove deprecated --disable-html-check option
+
+### Feature
+- Option to auto relay for matching recipient expression only ([#274](https://github.com/axllent/mailpit/issues/274))
+- Add UI settings screen
+
+### Fix
+- Add delay to close database on fatal exit ([#280](https://github.com/axllent/mailpit/issues/280))
+
+
 ## [v1.16.0]
 
 ### Chore